refactor: split sqlite_db file to match validator-node branch
